+++ scripts/macros/2024/spells/flameBlade.js
@@ -1,0 +1,6 @@
+import {flameBlade as flameBladeLegacy} from '../../../legacyMacros.js';
+export let flameBlade = {
+    ...flameBladeLegacy,
+    version: '1.2.24',
+    rules: 'modern'
+};